# docs/api_spec.yaml
components:
  schemas:
...
    ModuleListResponse:
      properties:
        modules:
          items:
            $ref: '#/components/schemas/ModuleResponse'
          type: array
        total:
          example: 5
          type: integer
      required:
      - modules
      - total
      type: object
    ModuleRegistration:
      properties:
        endpoint:
          example: http://localhost:9001
          format: uri
          type: string
        name:
          example: analytics-module
          type: string
        version:
          example: 1.2.0
          type: string
      required:
      - name
      - version
      - endpoint
      type: object
    ModuleResponse:
      properties:
        module_id:
          example: mod_123456
          type: string
        name:
          example: analytics-module
          type: string
        status:
          example: active
          type: string
        version:
          example: 1.2.0
          type: string
      required:
      - module_id
      - name
      - version
      - status
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: support@liftos.com
    name: Lift OS Core Team
  description: Unified API documentation for the Lift OS Core platform - a modular,
    scalable operating system for orchestrating all Lift products
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Lift OS Core API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /registry/modules:
    get:
      description: Get list of all registered modules
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleListResponse'
          description: Modules retrieved successfully
        '401':
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: List Modules
      tags:
      - Module Management
    post:
      description: Register a new module in the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleRegistration'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleResponse'
          description: Module registered successfully
        '400':
          description: Invalid module data
        '401':
          description: Unauthorized
        '409':
          description: Module already exists
      security:
      - BearerAuth: []
      summary: Register Module
      tags:
      - Module Management
servers:
- description: Development Gateway Server
  url: http://localhost:8000
- description: Production Gateway Server
  url: https://api.liftos.com